Defer `$botman->listen()` to the BotManController
Require that the `$botman->listen()` command come through the controller so that any middleware that is added in the Controller can be tested. 
Alternatively, the middleware would need to be added outside of the BotManController. With the current "Place your BotMan logic here." comment, it seems like the controller is the right place for this logic.